feat: move core types into utils
